moved files from backend-ops to libs/devops
