--- apps/backend-ops/src/test-setup.ts
@@ -1,1 +1,0 @@
-﻿import 'jest';
