+++ libs/devops/src/index.ts
@@ -1,1 +1,16 @@
-export * from './lib/devops';
+export * from './lib/admin';
+export * from './lib/algolia';
+export * from './lib/backup';
+export * from './lib/clean-backups';
+export * from './lib/db-cleaning';
+export * from './lib/db-shrink';
+export * from './lib/emulator';
+export * from './lib/firebaseSetup';
+export * from './lib/generate-fixtures';
+export * from './lib/health-check';
+export * from './lib/migrations';
+export * from './lib/rescueJWP';
+export * from './lib/select-environment';
+export * from './lib/storage-cleaning';
+export * from './lib/tools';
+export * from './lib/users';
