--- libs/devops/src/lib/devops.ts
@@ -1,3 +1,0 @@
-export function devops(): string {
-  return 'devops';
-}
