+++ libs/devops/src/lib/firebaseSetup.ts
